Centralize the count of $locales_array
Seeing that this variable is `count`ed multiple times, especially in
initial testing for use, seems to make sense to just count the array
once and then use the value of that count where needed.

There might be other variables that could benefit from this; however,
the correction of this PR was all visible within a few lines of code.

# Conflicts:
#	files/includes/templates/YOUR_TEMPLATE/jscript/jscript_plugin_structured_data.php